requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-oauth-toolkit==2.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
jwcrypto==1.5.6
oauthlib==3.2.2
pillow==10.4.0
psycopg==3.2.1
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
redis==5.0.7
requests==2.32.3
sqlparse==0.5.0
stripe==10.3.0
typing_extensions==4.12.2
urllib3==2.2.2
